Dokumentasi Perubahan & Perbaikan Proyek MealNa 29 November 2025
Dokumen ini merangkum serangkaian perbaikan dan peningkatan signifikan yang dilakukan pada aplikasi MealNa, mulai dari perbaikan error kritis hingga peningkatan pengalaman pengguna (UX) dan antarmuka (UI).

1.	Perbaikan Error Kritis PERMISSION_DENIED pada Firebase Firestore
•Masalah: Aplikasi tidak dapat menyimpan data dari halaman "Atur Profil Anda". Muncul error PERMISSION_DENIED: Missing or insufficient permissions, yang menandakan data ditolak oleh server Firebase.

•Penyebab: Aturan Keamanan (Security Rules) default pada database Firebase Firestore terlalu ketat dan tidak mengizinkan operasi tulis (write) dari aplikasi.

•Solusi: Aturan Keamanan di Firebase Console diperbarui untuk memberikan izin tulis kepada pengguna yang telah terautentikasi (login) ke dokumen mereka sendiri.

•Aturan Baru yang Diterapkan:

Javarules_version = '2';
service cloud.firestore {
  match /databases/{database}/documents {
    match /users/{userId} {
      allow read, write: if request.auth != null && request.auth.uid == userId;
    }
  }
}

•Hasil: Masalah PERMISSION_DENIED berhasil diatasi, dan aplikasi kini dapat menyimpan data profil pengguna ke koleksi users di Firestore.

2.	Perubahan Alur Navigasi Aplikasi
•Permintaan:
Mengubah alur setelah pengguna menyimpan profil. Pengguna harus diarahkan ke halaman orientasi (onboarding) terlebih dahulu, bukan langsung ke halaman utama.

•Solusi:
Logika navigasi di dalam ProfileSetupActivity.kt diubah.

•File yang Dimodifikasi:
app/src/main/java/com/example/mealna/ProfileSetupActivity.kt

•Perubahan Spesifik:
Di dalam blok .addOnSuccessListener, target Intent diubah dari HomeActivity::class.java menjadi OnboardingActivity::class.java.3. Peningkatan Desain Visual Splash Screen

•Permintaan:
Membuat tampilan splash.xml menjadi lebih "keren dan menarik".•Solusi: Mengganti latar belakang polos dengan efek gradien radial gelap yang modern untuk memberikan kesan visual yang lebih premium.

•File yang Dibuat:
◦app/src/main/res/drawable/bg_splash_gradient.xml: Berisi definisi kode untuk gradien radial.

•File yang Dimodifikasi:
◦app/src/main/res/drawable/splash.xml: Diubah menjadi layer-list untuk menempatkan logo mealna_logo di atas lapisan latar belakang gradien yang baru.

3.	Rombak Total Halaman Orientasi (Onboarding)
•Permintaan:
Menambahkan tombol navigasi "Selanjutnya", indikator halaman, animasi, dan memperbaiki tata letak agar lebih rapi dan fungsional

.•Solusi:
Melakukan perombakan komprehensif pada fungsionalitas dan UI/UX halaman orientasi.

•Perubahan Fungsional & Struktural:
◦Navigasi: Tombol "Selanjutnya" dan "Lewati" ditambahkan ke layout utama.
◦Indikator Halaman:
Indikator titik (DotsIndicator) ditambahkan untuk menunjukkan posisi halaman aktif.
◦Logika Kontrol:
OnboardingActivity.kt dirombak total untuk mengelola ViewPager2, tombol navigasi, dan indikator. Tombol "Selanjutnya" kini secara dinamis berubah menjadi "Mulai" pada halaman terakhir.
◦Layout:
activity_onboarding.xml diubah menjadi layout utama yang menampung ViewPager dan semua kontrol navigasi. Layout per halaman (onboarding_1.xml, _2.xml, _3.xml) disederhanakan untuk hanya menampilkan konten.

•Perubahan Tampilan & Animasi:
◦Animasi:
File slide_in_bottom.xml dibuat untuk memberikan efek transisi yang halus saat layar orientasi pertama kali muncul.
◦Tata Letak: Semua elemen (gambar, judul, deskripsi, tombol) ditata ulang menggunakan ConstraintLayout untuk memastikan tampilan yang seimbang dan rapi di berbagai ukuran layar.

•Perubahan Dependensi Proyek:
◦Library com.tbuonomo:
dotsindicator:4.3 ditambahkan ke file build.gradle.kts (:app) untuk menyediakan fungsionalitas indikator titik.
◦Proses Gradle Sync dijalankan untuk mengunduh dan mengintegrasikan dependensi baru ke dalam proyek.

+++ app/src/main/java/com/example/mealna/classes/User.kt
@@ -10,7 +10,7 @@ data class User(
 
 data class UserProfile(
     val height: Int = 0,
-    val weight: Int = 0,
+    val weight: Double = 0.0,
     val age: Int = 0,
     val gender: String = "",
     val diseases: List<String> = emptyList(),
